# auth/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: Auth API
    version: 0.0.1
paths:
    /auth/refresh:
        post:
            tags:
                - Auth
            operationId: Auth_RefreshToken
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/RefreshTokenRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/RefreshTokenReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /auth/sign-in:
        post:
            tags:
                - Auth
            operationId: Auth_SignIn
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/SignInRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SignInReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /auth/sign-up:
        post:
            tags:
                - Auth
            operationId: Auth_SignUp
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/SignUpRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SignUpReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
components:
    schemas:
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
        RefreshTokenReply:
            type: object
            properties:
                tokens:
                    $ref: '#/components/schemas/Tokens'
        RefreshTokenRequest:
            type: object
            properties:
                refreshToken:
                    type: string
        SignInReply:
            type: object
            properties:
                tokens:
                    $ref: '#/components/schemas/Tokens'
        SignInRequest:
            type: object
            properties:
                username:
                    type: string
                password:
                    type: string
        SignUpReply:
            type: object
            properties:
                tokens:
                    $ref: '#/components/schemas/Tokens'
        SignUpRequest:
            type: object
            properties:
                username:
                    type: string
                title:
                    type: string
                gender:
                    type: string
                email:
                    type: string
                password:
                    type: string
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
        Tokens:
            type: object
            properties:
                accessToken:
                    type: string
                refreshToken:
                    type: string
tags:
    - name: Auth